requirements.txt
dill==0.3.6
einops==0.5.0
easydict==1.13
ftfy==6.1.1
idna==3.4
multiprocess==0.70.14
numpy==1.23.3
opencv-python==4.6.0.66
pandas==1.5.0
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-toolbelt==0.10.1
scipy==1.9.2
six==1.16.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
EasyFL==0.1.2
